Create a SubjectProxy and separate CommandCause from CommandContext
This actually doesn't change much, but it was actually a problem in our implementation. It makes sense to keep it separate, and all the standard tasks can still be done in the same way (the context is now a SubjectProxy and you can easily send a message from it)
